# Conflicts:
#	src/main/java/pe/edu/upc/ecohabitproyecto/repositories/ITransaccionRepository.java
#	src/main/java/pe/edu/upc/ecohabitproyecto/servicesimplements/TransaccionServiceImplement.java
#	src/main/java/pe/edu/upc/ecohabitproyecto/servicesinterfaces/ITransaccionService.java

+++ src/main/java/pe/edu/upc/ecohabitproyecto/servicesinterfaces/IParticipacionDesafioService.java
@@ -5,11 +5,9 @@ import pe.edu.upc.ecohabitproyecto.entities.ParticipacionDesafio;
 import java.util.List;
 
 public interface IParticipacionDesafioService {
-
     List<ParticipacionDesafio> list();
     void insert(ParticipacionDesafio participacionDesafio);
-    ParticipacionDesafio listId(long id);
-    void delete(long id);
-
+    ParticipacionDesafio listId(Integer id);
+    void delete(Integer id);
     void update(ParticipacionDesafio participacionDesafio);
 }
